Change dependency structure for themes and devices
